Move logout function to profile page
